test: put every test that leaves the runner behind one feature
`#[ignore]` was carrying four unrelated meanings in this tree: needs the
corporate network, pins a known regression, regenerates a fixture, flakes
under load. A lane could only say yes or no to all four, so the previous
commit said no — which also switched off regression pins that had nothing
to do with the network.

The tests that reach a host this workspace does not own now live in
`suite_network`, gated by `required-features = ["network"]`. A default
`just test` does not compile them, no lane enables the feature, and no
`--run-ignored` can reach them. `#[ignore]` goes back to meaning what it
says on the remaining cases. Run the lane by hand from somewhere with the
network and the credentials: `just test run --lane=network`.

Two files mixed local and remote tests, so they split along that line:
`resource_regressions.rs` keeps its fixture-backed cases and hands the two
live-stream ones to `live_remote_network.rs`, and the user-simulation
suite keeps its fixture scenarios while the production ones move to
`prod_network.rs`.

That split left the shared scenario machinery serving two binaries, each
using a different part of it — dead code in whichever one used less. It
belongs in the test library, so `user_sim/{actions,harness,scenarios}` and
`test_defaults` move to `tests/src`. `common/mod.rs` re-exports
`test_defaults` under its old path, so the twenty-four files that read it
are untouched.

The move also put that code under `clippy --workspace` for the first time
— test targets are not linted there — and it had six findings. Two match
arms that returned the same thing for a delivered event and a closed bus
now say so in one arm; `STAGNATION_TICKS` was declared twice with the same
value inside two function bodies and is now one module constant beside
`POLL_INTERVAL`; `SawWav::default` is the trait it was imitating. Nothing
is suppressed. `SimHarness::current_variant` was a private delegate nobody
called and is gone.

86 tests now list under `suite_network`.
