+++ .config/xtask.toml
@@ -189,11 +189,9 @@ default_flash = false
 # on Rubato; inert off Apple), `e2e-glide` the Glide backend (no Rubato in the
 # graph, otherwise the compile-time alias picks Rubato).
 #
-# None of them passes `--run-ignored`. The tests that reach zvuk staging and
-# production say so themselves with `#[ignore]`, and a lane that overrides that
-# reports the fence around the runners rather than anything about the code. Run
-# them by hand with `--run-ignored=only` from a machine that has the network and
-# the credentials.
+# All three run `suite_e2e`, which reaches nothing outside the runner. The tests
+# that need the zvuk CDN or the corporate VPN live in `suite_network` behind the
+# `network` feature, so no amount of `--run-ignored` can pull them in here.
 [test.lanes.e2e]
 program = "cargo"
 prefix_args = [
@@ -237,6 +235,26 @@ prefix_args = [
   "suite_e2e",
 ]
 default_features = ["e2e", "kithara/resample-glide"]
+default_flash = false
+
+# Everything that talks to a host this workspace does not own: the zvuk
+# production CDN behind `KITHARA_DRM_PROD_*`, and `*.zvq.me`, which resolves
+# only inside the corporate VPN. Deliberately wired to no CI job — the runners
+# sit outside both networks, and a lane that runs there measures the fence, not
+# the code. Run it by hand: `just test run --lane=network`.
+[test.lanes.network]
+program = "cargo"
+prefix_args = [
+  "nextest",
+  "run",
+  "-p",
+  "kithara-integration-tests",
+  "--cargo-profile",
+  "test-release",
+  "--test",
+  "suite_network",
+]
+default_features = ["network"]
 default_flash = false
 
 # Integration regression tests are red by design and stay outside the main gate.
