+++ tests/src/test_defaults.rs
@@ -1,11 +1,12 @@
 use kithara::platform::{sync::Arc, time::Duration};
-use kithara_integration_tests::{audio_fixture::EmbeddedAudio, wav::create_test_wav};
+
+use crate::{audio_fixture::EmbeddedAudio, wav::create_test_wav};
 
 /// Default audio parameters for generated WAV test fixtures.
 ///
 /// All stress and integration tests that create synthetic WAV data
 /// use these values unless a test-specific override is needed.
-pub(crate) struct SawWav {
+pub struct SawWav {
     pub sample_rate: u32,
     pub channels: u16,
     pub segment_size: usize,
@@ -18,10 +19,10 @@ impl SawWav {
     const BYTES_PER_SAMPLE: usize = 2;
 
     /// Saw period in PCM frames (matches generated saw-tooth cycle length).
-    pub(crate) const SAW_PERIOD: usize = 65536;
+    pub const SAW_PERIOD: usize = 65536;
 
     /// Standard defaults: 44.1 kHz stereo, 200 KB segments (native) / 32 KB (wasm).
-    pub(crate) const DEFAULT: Self = Self {
+    pub const DEFAULT: Self = Self {
         sample_rate: 44100,
         channels: 2,
         #[cfg(not(target_arch = "wasm32"))]
@@ -30,15 +31,11 @@ impl SawWav {
         segment_size: 32_000,
     };
 
-    pub(crate) fn default() -> Self {
-        Self::DEFAULT
-    }
-
     /// Duration of one `segment_size`-byte slice in seconds.
     ///
     /// `bytes / (sample_rate * channels * bytes_per_sample)`. The same
     /// expression is inlined in several HLS fixture builders.
-    pub(crate) fn segment_duration_secs(&self) -> f64 {
+    pub fn segment_duration_secs(&self) -> f64 {
         self.segment_size as f64
             / (f64::from(self.sample_rate)
                 * f64::from(self.channels)
@@ -46,7 +43,7 @@ impl SawWav {
     }
 
     /// Byte count for `segments` consecutive segments of `segment_size`.
-    pub(crate) fn total_bytes(&self, segments: usize) -> usize {
+    pub fn total_bytes(&self, segments: usize) -> usize {
         segments * self.segment_size
     }
 
@@ -54,11 +51,17 @@ impl SawWav {
     ///
     /// Uses [`create_test_wav`] with a sine wave; frame count is derived so
     /// the header + PCM body fit within the total byte budget.
-    pub(crate) fn build_wav(&self, segments: usize) -> Arc<Vec<u8>> {
+    pub fn build_wav(&self, segments: usize) -> Arc<Vec<u8>> {
         let total = self.total_bytes(segments);
         let bytes_per_frame = self.channels as usize * Self::BYTES_PER_SAMPLE;
         let frames = (total - Self::WAV_HEADER_SIZE) / bytes_per_frame;
         Arc::new(create_test_wav(frames, self.sample_rate, self.channels))
+    }
+}
+
+impl Default for SawWav {
+    fn default() -> Self {
+        Self::DEFAULT
     }
 }
 
@@ -72,23 +75,23 @@ impl SawWav {
 ///
 /// New entries should be added only when the same value recurs in ≥2
 /// modules. Genuinely local values stay in their own file.
-pub(crate) struct Consts;
+pub struct Consts;
 
 impl Consts {
     /// Default sample rate for generated WAV / expected streams.
-    pub(crate) const SAMPLE_RATE: u32 = SawWav::DEFAULT.sample_rate;
+    pub const SAMPLE_RATE: u32 = SawWav::DEFAULT.sample_rate;
     /// Default channel count.
-    pub(crate) const CHANNELS: u16 = SawWav::DEFAULT.channels;
+    pub const CHANNELS: u16 = SawWav::DEFAULT.channels;
     /// Default packaged HLS segment size (bytes).
-    pub(crate) const SEGMENT_SIZE: usize = SawWav::DEFAULT.segment_size;
+    pub const SEGMENT_SIZE: usize = SawWav::DEFAULT.segment_size;
 
     /// Bytes of the `assets/test.mp3` fixture.
-    pub(crate) const TEST_MP3_BYTES: &'static [u8] = EmbeddedAudio::TEST_MP3_BYTES;
+    pub const TEST_MP3_BYTES: &'static [u8] = EmbeddedAudio::TEST_MP3_BYTES;
     /// Nominal duration of [`Self::TEST_MP3_BYTES`] in seconds.
-    pub(crate) const TEST_MP3_DURATION_SECS: f64 = EmbeddedAudio::MP3_EXPECTED_DURATION_SECS;
+    pub const TEST_MP3_DURATION_SECS: f64 = EmbeddedAudio::MP3_EXPECTED_DURATION_SECS;
 
     /// Default soft read timeout for resource/decoder integration tests.
-    pub(crate) const READ_TIMEOUT: Duration = Duration::from_secs(5);
+    pub const READ_TIMEOUT: Duration = Duration::from_secs(5);
     /// Block size used by offline render harnesses (audio block budget).
-    pub(crate) const OFFLINE_BLOCK_FRAMES: usize = 512;
+    pub const OFFLINE_BLOCK_FRAMES: usize = 512;
 }
