+++ tests/src/user_sim/mod.rs
@@ -1,0 +1,3 @@
+pub mod actions;
+pub mod harness;
+pub mod scenarios;
