+++ tests/tests/common/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod test_defaults;
+pub(crate) use kithara_integration_tests::test_defaults;
