+++ tests/tests/kithara_play/resource_regressions.rs
@@ -9,17 +9,13 @@ use kithara::{
     decode::DecoderBackend,
     file::{File as FileSource, FileConfig, FileSrc},
     hls::{Hls, HlsConfig},
-    net::{HttpClient, NetOptions},
     platform::{
         CancelToken,
         sync::Arc,
         time::{Duration, Instant, sleep, timeout},
     },
     play::{PlayerConfig, PlayerImpl, Resource, ResourceConfig},
-    stream::{
-        AudioCodec, ContainerFormat, MediaInfo, Stream,
-        dl::{Downloader, DownloaderConfig},
-    },
+    stream::{AudioCodec, ContainerFormat, MediaInfo, Stream},
 };
 use kithara_integration_tests::{
     Content, Delivery, FixtureBehavior, HlsFixtureBuilder, TestServerHelper, TestTempDir,
@@ -30,7 +26,7 @@ use kithara_integration_tests::{
     signal_pcm::signal,
     temp_dir,
 };
-use tracing::{debug, info};
+use tracing::info;
 
 use crate::{
     common::test_defaults::Consts as Shared,
@@ -1321,298 +1317,11 @@ async fn resource_mp3_no_hint_decodes_with_duration(
     );
 }
 
-/// Live remote streams through `ResourceConfig` — same code path as kithara-app.
-/// No hint, no extension manipulation — exactly what the app does.
-///
-/// Requires internet (silvercomet) and corporate VPN (zvuk).
-// flash(false): live-internet sockets are invisible to the flash engine; virtual
-// sleep/deadline would outrun the real download and fail spuriously.
-#[kithara::test(
-    tokio,
-    timeout(Duration::from_secs(30)),
-    env(
-        KITHARA_HANG_TIMEOUT_SECS = "10",
-        http_proxy = "",
-        https_proxy = "",
-        HTTP_PROXY = "",
-        HTTPS_PROXY = ""
-    )
-)]
-#[ignore = "requires internet; zvuk URLs require corporate VPN"]
-#[case::silvercomet_mp3_symphonia(
-    "https://stream.silvercomet.top/track.mp3",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::silvercomet_mp3_apple("https://stream.silvercomet.top/track.mp3", DecoderBackend::Apple)
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::silvercomet_mp3_android(
-        "https://stream.silvercomet.top/track.mp3",
-        DecoderBackend::Android
-    )
-)]
-#[case::silvercomet_hls_symphonia(
-    "https://stream.silvercomet.top/hls/master.m3u8",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::silvercomet_hls_apple(
-        "https://stream.silvercomet.top/hls/master.m3u8",
-        DecoderBackend::Apple
-    )
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::silvercomet_hls_android(
-        "https://stream.silvercomet.top/hls/master.m3u8",
-        DecoderBackend::Android
-    )
-)]
-#[case::zvuk_27390231_symphonia(
-    "https://cdn-edge.zvq.me/track/streamhq?id=27390231",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::zvuk_27390231_apple(
-        "https://cdn-edge.zvq.me/track/streamhq?id=27390231",
-        DecoderBackend::Apple
-    )
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::zvuk_27390231_android(
-        "https://cdn-edge.zvq.me/track/streamhq?id=27390231",
-        DecoderBackend::Android
-    )
-)]
-#[case::zvuk_151585912_symphonia(
-    "https://cdn-edge.zvq.me/track/streamhq?id=151585912",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::zvuk_151585912_apple(
-        "https://cdn-edge.zvq.me/track/streamhq?id=151585912",
-        DecoderBackend::Apple
-    )
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::zvuk_151585912_android(
-        "https://cdn-edge.zvq.me/track/streamhq?id=151585912",
-        DecoderBackend::Android
-    )
-)]
-#[case::zvuk_125475417_symphonia(
-    "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::zvuk_125475417_apple(
-        "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-        DecoderBackend::Apple
-    )
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::zvuk_125475417_android(
-        "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-        DecoderBackend::Android
-    )
-)]
-async fn live_remote_resource_decodes_with_duration(
-    #[case] url: &str,
-    #[case] backend: DecoderBackend,
-    temp_dir: TestTempDir,
-) {
-    #[cfg(any(target_os = "macos", target_os = "ios"))]
-    kithara_integration_tests::apple_warmup::warm_if_apple(backend);
-
-    let store = asset_store(&temp_dir, true);
-    let net = NetOptions::builder()
-        .inactivity_timeout(Duration::from_secs(25))
-        .build();
-    let downloader = Downloader::new(
-        DownloaderConfig::for_client(HttpClient::new(net, CancelToken::never())).build(),
-    );
-    let config: ResourceConfig =
-        ResourceConfig::for_src(ResourceConfig::parse_src(url).expect("valid URL"))
-            .byte_pool(kithara::bufpool::BytePool::default())
-            .pcm_pool(kithara::bufpool::PcmPool::default())
-            .store(store)
-            .downloader(downloader)
-            .decoder(
-                kithara::audio::AudioDecoderConfig::builder()
-                    .backend(backend)
-                    .build(),
-            )
-            .build();
-
-    let mut resource = Resource::new(config)
-        .await
-        .unwrap_or_else(|e| panic!("{url}: Resource::new failed: {e}"));
-
-    let duration = resource.duration();
-    assert!(
-        duration.is_some(),
-        "{url}: duration must be reported (got None)"
-    );
-    let dur_secs = duration.expect("checked").as_secs_f64();
-    assert!(
-        dur_secs > 30.0,
-        "{url}: expected duration > 30s for a real track, got {dur_secs:.1}s"
-    );
-
-    let deadline = Instant::now() + Duration::from_secs(20);
-    let mut samples = 0usize;
-    let mut buf = [0.0f32; 4096];
-    loop {
-        match resource.read(&mut buf) {
-            Ok(ReadOutcome::Frames { count, .. }) => {
-                let count = count.get();
-                if count > 0 {
-                    samples += count;
-                }
-            }
-            Ok(ReadOutcome::Eof { .. }) => break,
-            Ok(ReadOutcome::Pending { .. }) => {}
-            Err(e) => panic!("{url}: decode error: {e}"),
-        }
-        if resource.position() >= Duration::from_secs(2) {
-            break;
-        }
-        assert!(
-            Instant::now() <= deadline,
-            "{url}: timed out waiting for PCM data (pos={:?}, samples={samples})",
-            resource.position()
-        );
-        time::sleep(Duration::from_millis(5)).await;
-    }
-
-    assert!(samples > 0, "{url}: must decode PCM samples");
-    assert!(
-        resource.position() >= Duration::from_secs(2),
-        "{url}: must decode at least 2s, got {:?}",
-        resource.position()
-    );
-}
-
-/// Reproduces EXACTLY the app flow: `PlayerImpl` + `prepare_config` + `Resource::new` +
-/// `select_item` + `duration_seconds()`. This is what the GUI reads.
-// flash(false): live-internet sockets are invisible to the flash engine; a virtual
-// 500ms pacing sleep would elapse before the real metadata fetch completes.
-#[kithara::test(
-    tokio,
-    timeout(Duration::from_secs(30)),
-    env(
-        KITHARA_HANG_TIMEOUT_SECS = "10",
-        http_proxy = "",
-        https_proxy = "",
-        HTTP_PROXY = "",
-        HTTPS_PROXY = ""
-    )
-)]
-#[ignore = "requires internet; zvuk URLs require corporate VPN"]
-#[case::silvercomet_mp3_symphonia(
-    "https://stream.silvercomet.top/track.mp3",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::silvercomet_mp3_apple("https://stream.silvercomet.top/track.mp3", DecoderBackend::Apple)
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::silvercomet_mp3_android(
-        "https://stream.silvercomet.top/track.mp3",
-        DecoderBackend::Android
-    )
-)]
-#[case::zvuk_125475417_symphonia(
-    "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-    DecoderBackend::Symphonia
-)]
-#[cfg_attr(
-    any(target_os = "macos", target_os = "ios"),
-    case::zvuk_125475417_apple(
-        "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-        DecoderBackend::Apple
-    )
-)]
-#[cfg_attr(
-    target_os = "android",
-    case::zvuk_125475417_android(
-        "https://cdn-edge.zvq.me/track/streamhq?id=125475417",
-        DecoderBackend::Android
-    )
-)]
-async fn player_mp3_duration_matches_app_flow(
-    #[case] url: &str,
-    #[case] backend: DecoderBackend,
-    temp_dir: TestTempDir,
-) {
-    #[cfg(any(target_os = "macos", target_os = "ios"))]
-    kithara_integration_tests::apple_warmup::warm_if_apple(backend);
-
-    let store = asset_store(&temp_dir, true);
-
-    let player = PlayerImpl::new(
-        PlayerConfig::builder()
-            .byte_pool(kithara::bufpool::BytePool::default())
-            .pcm_pool(kithara::bufpool::PcmPool::default())
-            .build(),
-    );
-    player.reserve_slots(1);
-
-    let mut config = ResourceConfig::for_src(ResourceConfig::parse_src(url).unwrap())
-        .byte_pool(kithara::bufpool::BytePool::default())
-        .pcm_pool(kithara::bufpool::PcmPool::default())
-        .store(store)
-        .decoder(
-            kithara::audio::AudioDecoderConfig::builder()
-                .backend(backend)
-                .build(),
-        )
-        .build();
-    config = player.prepare_config(config);
-
-    let resource = Resource::new(config)
-        .await
-        .unwrap_or_else(|e| panic!("{url}: Resource::new failed: {e}"));
-
-    player.replace_item(0, resource);
-    let _ = player.select_item(0, true);
-
-    // Wait on the concrete state the assertion below reads: the selected
-    // slot's duration committed into player shared state. The inner sleep is
-    // only the poll cadence of this state-checking loop; the test-level
-    // `timeout(30s)` bounds a genuine stall.
-    while player.duration_seconds().is_none() {
-        time::sleep(Duration::from_millis(20)).await;
-    }
-
-    let dur = player.duration_seconds();
-    debug!("{url} duration_seconds={dur:?}");
-    assert!(
-        dur.is_some(),
-        "{url}: player.duration_seconds() must not be None"
-    );
-    let dur_secs = dur.expect("checked");
-    assert!(dur_secs > 30.0, "{url}: expected >30s, got {dur_secs:.1}s");
-
-    player.worker().shutdown();
-}
-
 /// Local fixture (`assets/track.mp3` ~162s, packaged AAC HLS ~64s) through
 /// `ResourceConfig` — same code path as kithara-app. Mirrors
-/// `live_remote_resource_decodes_with_duration` but against `TestServerHelper`
-/// so it stays in the regular suite (no `#[ignore]`, no VPN, no internet).
+/// `live_remote_resource_decodes_with_duration` (now in
+/// `live_remote_network.rs`) but against `TestServerHelper`, so it stays in the
+/// regular suite: no VPN, no internet.
 #[derive(Clone, Copy, Debug)]
 enum LocalKind {
     Mp3,
