+++ tests/tests/kithara_queue/user_simulation/mod.rs
@@ -1,8 +1,6 @@
 #![cfg(not(target_arch = "wasm32"))]
 #![forbid(unsafe_code)]
 
-pub(crate) mod actions;
-pub(crate) mod harness;
-pub(crate) mod scenarios;
+pub(crate) use kithara_integration_tests::user_sim::{actions, harness, scenarios};
 
 mod tests;
